Fix four correctness issues in ray tracing and shader module APIs
- wgpuCommandEncoderBuildAccelerationStructures: panic on unknown
  WGPUBlasGeometryKind instead of silently treating it as AABBs,
  matching the existing behavior in wgpuDeviceCreateBlas

- wgpuBlasPrepareCompactAsync: always invoke blas_prepare_compact_async
  even when callback is null; previously the function returned early
  without preparing the BLAS for compaction at all

- wgpuShaderModuleGetCompilationInfo: heap-allocate message strings via
  str_into_owned_string_view so they have clear ownership independent of
  the shader module, and free them after the callback returns

- WGPUNativeLimits ray tracing limits (maxBlasPrimitiveCount,
  maxBlasGeometryCount, maxTlasInstanceCount): change from uint64_t to
  uint32_t to match the underlying wgpu u32 types; removes silent
  truncation on the read path and a widening cast on the write path

+++ ffi/wgpu.h
@@ -1389,9 +1389,9 @@ typedef struct WGPUNativeLimits
     uint32_t maxMeshOutputLayers;
     uint32_t maxMeshMultiviewViewCount;
     /* Ray tracing limits */
-    uint64_t maxBlasPrimitiveCount;
-    uint64_t maxBlasGeometryCount;
-    uint64_t maxTlasInstanceCount;
+    uint32_t maxBlasPrimitiveCount;
+    uint32_t maxBlasGeometryCount;
+    uint32_t maxTlasInstanceCount;
     uint32_t maxAccelerationStructuresPerShaderStage;
 } WGPUNativeLimits;
